Reduce the number of API requests after scanning a barcode
The script sent a request everytime a barcode was detected,
which could trigger the rate limiter.
This is a "hacky" fix that sends a request every 10 detections.
The first one should be enough, but just in case the first request
wouldn't be served, it is allowed to send other ones.